# internal/controller/handler/v1/purchases/detail/pay/gen/server.gen.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package gen

import (
	"bytes"
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/labstack/echo/v5"
	"github.com/oapi-codegen/runtime"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// PatchPurchasesPay 購入の支払い
	// (PATCH /v1/purchases/{purchaseCode}/pay)
	PatchPurchasesPay(ctx *echo.Context, purchaseCode PurchaseCodeParam) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// PatchPurchasesPay converts echo context to params.
func (w *ServerInterfaceWrapper) PatchPurchasesPay(ctx *echo.Context) error {
	var err error
	// ------------- Path parameter "purchaseCode" -------------
	var purchaseCode PurchaseCodeParam

	err = runtime.BindStyledParameterWithOptions("simple", "purchaseCode", ctx.Param("purchaseCode"), &purchaseCode, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "", ValueIsUnescaped: ctx.Request().URL.RawPath == ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter purchaseCode: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PatchPurchasesPay(ctx, purchaseCode)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
}

// RegisterHandlersOptions configures RegisterHandlersWithOptions.
type RegisterHandlersOptions struct {
	// BaseURL is prepended to every registered path so the API can be served
	// under a prefix.
	BaseURL string
	// OperationMiddlewares lets the caller attach per-operation middleware at
	// registration time. The map key is the OpenAPI `operationId` value as it
	// appears in the spec (the raw, un-normalized form). Operations that have
	// no entry are registered with no extra middleware. A nil map disables
	// per-operation middleware entirely.
	OperationMiddlewares map[string][]echo.MiddlewareFunc
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, RegisterHandlersOptions{})
}

// RegisterHandlersWithBaseURL registers handlers and prepends BaseURL to the
// paths so the API can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {
	RegisterHandlersWithOptions(router, si, RegisterHandlersOptions{BaseURL: baseURL})
}

// RegisterHandlersWithOptions registers handlers using the supplied options,
// including any per-operation middleware.
func RegisterHandlersWithOptions(router EchoRouter, si ServerInterface, options RegisterHandlersOptions) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.PATCH(options.BaseURL+"/v1/purchases/:purchaseCode/pay", wrapper.PatchPurchasesPay, options.OperationMiddlewares["PatchPurchasesPay"]...)

}

type BadRequest400JSONResponse ErrorResponse

type Conflict409JSONResponse ErrorResponse

type InternalServerError500JSONResponse ErrorResponse

type MethodNotAllowed405JSONResponse ErrorResponse

type NotFound404JSONResponse ErrorResponse

type ServiceUnavailable503JSONResponse ErrorResponse

type Unauthorized401JSONResponse ErrorResponse

type PatchPurchasesPayRequestObject struct {
	PurchaseCode PurchaseCodeParam `json:"purchaseCode"`
}

type PatchPurchasesPayResponseObject interface {
	VisitPatchPurchasesPayResponse(w http.ResponseWriter) error
}

type PatchPurchasesPay200JSONResponse PurchasePayResponse

func (response PatchPurchasesPay200JSONResponse) VisitPatchPurchasesPayResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type PatchPurchasesPay400JSONResponse struct{ BadRequest400JSONResponse }

func (response PatchPurchasesPay400JSONResponse) VisitPatchPurchasesPayResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)
	_, err := buf.WriteTo(w)
	return err
}

type PatchPurchasesPay401JSONResponse struct{ Unauthorized401JSONResponse }

func (response PatchPurchasesPay401JSONResponse) VisitPatchPurchasesPayResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)
	_, err := buf.WriteTo(w)
	return err
}

type PatchPurchasesPay404JSONResponse struct{ NotFound404JSONResponse }

func (response PatchPurchasesPay404JSONResponse) VisitPatchPurchasesPayResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)
	_, err := buf.WriteTo(w)
	return err
}

type PatchPurchasesPay405JSONResponse struct {
	MethodNotAllowed405JSONResponse
}

func (response PatchPurchasesPay405JSONResponse) VisitPatchPurchasesPayResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(405)
	_, err := buf.WriteTo(w)
	return err
}

type PatchPurchasesPay409JSONResponse struct{ Conflict409JSONResponse }

func (response PatchPurchasesPay409JSONResponse) VisitPatchPurchasesPayResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(409)
	_, err := buf.WriteTo(w)
	return err
}

type PatchPurchasesPay500JSONResponse struct {
	InternalServerError500JSONResponse
}

func (response PatchPurchasesPay500JSONResponse) VisitPatchPurchasesPayResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)
	_, err := buf.WriteTo(w)
	return err
}

type PatchPurchasesPay503JSONResponse struct {
	ServiceUnavailable503JSONResponse
}

func (response PatchPurchasesPay503JSONResponse) VisitPatchPurchasesPayResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(503)
	_, err := buf.WriteTo(w)
	return err
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// PatchPurchasesPay 購入の支払い
	// (PATCH /v1/purchases/{purchaseCode}/pay)
	PatchPurchasesPay(ctx context.Context, request PatchPurchasesPayRequestObject) (PatchPurchasesPayResponseObject, error)
}

type StrictHandlerFunc func(ctx *echo.Context, request any) (any, error)
type StrictMiddlewareFunc func(f StrictHandlerFunc, operationID string) StrictHandlerFunc

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
}

// PatchPurchasesPay operation middleware
func (sh *strictHandler) PatchPurchasesPay(ctx *echo.Context, purchaseCode PurchaseCodeParam) error {
	var request PatchPurchasesPayRequestObject

	request.PurchaseCode = purchaseCode

	handler := func(ctx *echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.PatchPurchasesPay(ctx.Request().Context(), request.(PatchPurchasesPayRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PatchPurchasesPay")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(PatchPurchasesPayResponseObject); ok {
		return validResponse.VisitPatchPurchasesPayResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}