# openapi/openapi.gen.yaml
openapi: 3.0.3
info:
  title: Go Boilerplate API
  termsOfService: https://github.com/Tomy-ch/go-boilerplate-echo-oapi-sqlc
  contact:
    name: TomyC258460
    url: https://x.com/TomyC258460
  version: 2.2.0
  description: |
    Reference specification for the sample API (users, products, purchases). It exercises every layer
    of the Onion Architecture end to end: cursor and offset pagination, creation and partial update,
    soft delete, idempotency keys, and role-based authorization.

    Handlers (oapi-codegen) and the published reference documentation are both generated from this
    file, so every endpoint change starts here.
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
    x-internal: true
  - url: https://api.example.com
    description: Production server (replace with actual URL)
    x-internal: false
  - url: https://staging.api.example.com
    description: Staging server (replace with actual URL)
    x-internal: false
  - url: https://development.api.example.com
    description: Development server (replace with actual URL)
    x-internal: true
security: []
...
paths:
...
  /v1/products/ranking:
    get:
      operationId: GetProductsRanking
      tags:
        - v1/products/ranking
      summary: 商品売上ランキングの取得
      description: |
        購入明細を集計し、販売数量の多い順に商品ランキングを返します。認証不要の公開エンドポイントです。
        キャンセル済みの購入は集計から除外します。未払いの購入は集計に含みます。
        period=30d の場合は注文日時が直近30日以内の購入のみを集計します。
        同一販売数量の商品は商品 ID の昇順で安定的に並びます。
      security: []
      parameters:
        - $ref: '#/components/parameters/RankingPeriodParam'
        - $ref: '#/components/parameters/RankingLimitParam'
      responses:
        '200':
          description: 販売数量の降順で取得した商品ランキング
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductRankingResponse'
        '400':
          $ref: '#/components/responses/BadRequest400'
        '405':
          $ref: '#/components/responses/MethodNotAllowed405'
        '500':
          $ref: '#/components/responses/InternalServerError500'
        '503':
          $ref: '#/components/responses/ServiceUnavailable503'
...
  /v1/purchases/{purchaseCode}/pay:
    patch:
      operationId: PatchPurchasesPay
      tags:
        - v1/purchases/detail/pay
      summary: 購入の支払い
      description: |
        認証主体（本人）の購入を支払い済みへ遷移させます。認証必須です。
        決済 SDK / PSP 連携は行わず、paidAt のセットと status の「支払い済み」への更新で擬似決済を代替します
        （金額・決済結果の検証は行いません）。未払い相当（未処理 / 受付中 / 確認中 / 処理中）からのみ遷移でき、
        既に支払い済みの購入は 409（二重支払い）、キャンセル済み・完了・発送済み・配達済みの購入は 409（不正遷移）で拒否します。
        他ユーザーの購入は 404 で存在を秘匿します。金額はすべて USD セント単位の整数です。
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PurchaseCodeParam'
      responses:
        '200':
          description: 購入の支払いに成功しました
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchasePayResponse'
        '400':
          $ref: '#/components/responses/BadRequest400'
        '401':
          $ref: '#/components/responses/Unauthorized401'
        '404':
          $ref: '#/components/responses/NotFound404'
        '405':
          $ref: '#/components/responses/MethodNotAllowed405'
        '409':
          $ref: '#/components/responses/Conflict409'
        '500':
          $ref: '#/components/responses/InternalServerError500'
        '503':
          $ref: '#/components/responses/ServiceUnavailable503'
...
components:
  parameters:
...
    RankingPeriodParam:
      name: period
      in: query
      description: |
        集計対象期間。`all`（既定）は全期間、`30d` は注文日時が直近30日以内の購入のみを集計します。
      required: false
      schema:
        type: string
        enum:
          - all
          - 30d
        default: all
        example: all
    RankingLimitParam:
      name: limit
      in: query
      description: 取得する上位件数。売上数量の降順で上位 limit 件を返します。
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
        example: 10
...
    PurchaseCodeParam:
      name: purchaseCode
      in: path
      required: true
      description: 購入コード（利用者へ注文番号として見せる一意の識別子）
      schema:
        type: string
        minLength: 1
        maxLength: 50
        example: 0190b0d4-7b1a-7c2e-9f3a-1b2c3d4e5f60
...
  securitySchemes:
    BearerAuth:
      type: http
      description: |
        認証にBearerトークンを使用することを示します。
        JWT の検証は RFC8725（JSON Web Token Best Current Practices）に従います。
        署名アルゴリズムは許可リストで固定し、`alg: none` と鍵の混同を拒否します。
      scheme: bearer
      bearerFormat: JWT
    BasicAuth:
      type: http
      description: 認証にBasic認証を使用することを示します。
      scheme: basic
  schemas:
    ErrorResponse:
      type: object
      description: |
        エラーレスポンスの共通スキーマ（base）。 details は返さない。details を返すエンドポイントは ErrorResponseWithDetails を参照する。
      required:
        - code
        - message
        - requestId
      properties:
        code:
          type: string
          description: 機械的に処理可能なエラーコード
          example: EXAMPLE_ERROR
        message:
          type: string
          description: 人間が読めるエラーメッセージ
          example: 定義されたエラーメッセージ
        requestId:
          type: string
          description: リクエストID
          example: abcdefghijklmnopqrstuvwxyz
...
    ProductRankingItem:
      type: object
      description: 売上ランキング 1 商品分の項目です。
      required:
        - productId
        - name
        - price
        - soldQuantity
      properties:
        productId:
          type: string
          format: uuid
          description: 商品ID
          example: b1d4e0f2-3c5a-4b6d-8e7f-1a2b3c4d5e6f
        name:
          type: string
          description: 商品名
          maxLength: 255
          example: ワイヤレスイヤホン
        price:
          type: string
          pattern: ^\d+(\.\d+)?$
          description: 価格。サブセント精度を保持する decimal 文字列で表します（例 "19.99" は 19.99 ドル）。 JSON number は IEEE754 double として復元され精度を失うため、文字列で表現します。
          example: '19.99'
        soldQuantity:
          type: integer
          format: int64
          description: 集計期間内の販売数量（キャンセル済みの購入を除く合算値）。
          example: 42
    ProductRankingResponse:
      type: object
      description: 売上ランキングのレスポンススキーマ。販売数量の降順（同数量は商品 ID 昇順）で並びます。
      required:
        - rankings
      properties:
        rankings:
          type: array
          description: 販売数量の降順で並んだ商品ランキングです。
          items:
            $ref: '#/components/schemas/ProductRankingItem'
...
    PurchaseStatusRef:
      type: object
      description: |
        購入に紐づくステータス（ID・業務キー・名称）。name は事前解決済みで、別途の名称解決 API 呼び出しは不要です。
        表示には name を、分岐にはドメインの業務キーである code を用います。
      required:
        - id
        - code
        - name
      properties:
        id:
          type: string
          format: uuid
          description: 購入ステータスID
          example: 0190b0d4-7b1a-7c2e-9f3a-1b2c3d4e5f60
        code:
          type: integer
          format: int64
          description: 購入ステータスの業務キー。値は到達順序を意味しません（完了より支払い済みのほうが大きい）。
          example: 7
        name:
          type: string
          description: 購入ステータス名
          example: 完了
...
    PurchaseDetailResponse:
      type: object
      description: 購入明細のレスポンス。unitPrice は購入時点の価格スナップショットです。
      required:
        - productId
        - quantity
        - unitPrice
      properties:
        productId:
          type: string
          format: uuid
          description: 商品ID
          example: 5266f905-1af4-477a-bc7c-e69729e22a2b
        quantity:
          type: integer
          format: int32
          description: 購入数量
          example: 2
        unitPrice:
          type: string
          pattern: ^\d+(\.\d+)?$
          description: |
            購入時点の単価スナップショット。サブセント精度を保持する decimal 文字列（USD ドル）で表します（例 "800.00"）。
            購入成立後に商品価格が変わっても本値は不変です。
          example: '800.00'
...
    PurchasePayResponse:
      type: object
      description: |
        購入支払い後の購入情報レスポンススキーマ。status は ID・名称を事前解決済みで、
        別途の名称解決 API 呼び出しは不要です。paidAt に支払日時が入ります。
        金額（subtotalAmount / taxAmount / shippingFee / totalAmount / 明細 unitPrice）は
        すべて USD セント単位の整数です。
      required:
        - code
        - userId
        - status
        - subtotalAmount
        - taxAmount
        - shippingFee
        - totalAmount
        - details
        - orderedAt
        - paidAt
      properties:
        code:
          type: string
          description: 購入コード（利用者へ注文番号として見せる一意の識別子）
          example: 0190b0d4-7b1a-7c2e-9f3a-1b2c3d4e5f60
        userId:
          type: string
          format: uuid
          description: 購入したユーザーのID
          example: 550e8400-e29b-41d4-a716-446655440000
        status:
          $ref: '#/components/schemas/PurchaseStatusRef'
        subtotalAmount:
          type: integer
          format: int64
          description: 小計。明細（単価×数量）の合計。USD セント単位の整数です。
          example: 160000
        taxAmount:
          type: integer
          format: int64
          description: 税額（国内消費税）。USD セント単位の整数です。
          example: 16000
        shippingFee:
          type: integer
          format: int64
          description: 送料。USD セント単位の整数です。
          example: 500
        totalAmount:
          type: integer
          format: int64
          description: 合計（小計 + 税額 + 送料）。USD セント単位の整数です。
          example: 176500
        details:
          type: array
          description: 購入明細の配列
          items:
            $ref: '#/components/schemas/PurchaseDetailResponse'
        orderedAt:
          type: string
          format: date-time
          description: 注文日時
          example: '2026-07-23T00:00:00Z'
        paidAt:
          type: string
          format: date-time
          description: 支払日時
          example: '2026-07-25T00:00:00Z'
...
  responses:
    MethodNotAllowed405:
      description: 許可されていないリクエスト方法です。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequest400:
      description: リクエストが不正です。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized401:
      description: 認証が必要です。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError500:
      description: サーバーエラーが発生しました。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServiceUnavailable503:
      description: サービスが一時的に利用できません。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict409:
      description: リソースが競合しています。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound404:
      description: 指定されたリソースが存在しません。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...